Move Reference and ItemReference to model package
