+++ tests/util/test_parse_urn.py
@@ -1,7 +1,8 @@
 import pytest
 
 from pysdmx.errors import Invalid
-from pysdmx.util import Reference, parse_urn
+from pysdmx.model import Reference
+from pysdmx.util import parse_urn
 
 
 def test_no_match():
